--- src/main/java/com/triple/backend/auth/controller/LoginController.java
@@ -1,23 +1,0 @@
-package com.triple.backend.auth.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@RestController
-@RequiredArgsConstructor
-public class LoginController {
-
-    @GetMapping("/login")
-    public String loginPage() {
-        return "login"; // login.html 파일을 반환합니다. (뷰 리졸버에 따라 경로 조정 필요)
-    }
-
-}
