+++ src/main/resources/static/login.html
@@ -95,7 +95,7 @@
         const password = document.getElementById("password").value;
 
         // 로그인 요청 보내기
-        fetch("/login", {
+        fetch("/auth/login", {
             method: "POST",
             headers: {
                 "Content-Type": "application/x-www-form-urlencoded",
