+++ src/main/resources/static/signup.html
@@ -60,7 +60,7 @@
 </header>
 
 <main>
-    <form action="/join" method="post">
+    <form action="/auth/signup" method="post">
         <label for="memberName">이름</label>
         <input type="text" id="memberName" name="memberName" required>
         <br>
